feat: jitsu analytics plugin by NX generator
