+++ packages/plugins/integration-upwork/package.json
@@ -43,7 +43,7 @@
 		"access": "restricted"
 	},
 	"scripts": {
-		"test:e2e": "jest --config ./jest.config.js",
+		"test:e2e": "jest --config ./jest.config.ts",
 		"build": "rimraf dist && tsc -p tsconfig.lib.json",
 		"build:prod": "rimraf dist && tsc -p tsconfig.lib.prod.json"
 	},
