+++ packages/plugins/jitsu-analytics/.gitignore
@@ -1,0 +1,6 @@
+# dependencies
+node_modules/
+
+# misc
+npm-debug.log
+dist
