--- packages/plugins/jitsu-analytics/jest.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-	name: 'jitsu-analytics',
-	preset: '../../../jest.config.js',
-	transform: {
-		'^.+\\.[tj]sx?$': 'ts-jest'
-	},
-	moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
-	coverageDirectory: '../../coverage/plugins/jitsu-analytics'
-};
