+++ packages/plugins/jitsu-analytics/jest.config.ts
@@ -1,11 +1,11 @@
 /* eslint-disable */
 export default {
-	displayName: 'integration-github',
+	displayName: 'jitsu-analytics',
 	preset: '../../../jest.preset.js',
 	testEnvironment: 'node',
 	transform: {
 		'^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }]
 	},
 	moduleFileExtensions: ['ts', 'js', 'html'],
-	coverageDirectory: '../../../coverage/packages/plugins/integration-github'
+	coverageDirectory: '../../../coverage/packages/plugins/jitsu-analytics'
 };
