+++ packages/plugins/jitsu-analytics/package.json
@@ -38,9 +38,9 @@
 		"access": "restricted"
 	},
 	"scripts": {
-		"test:e2e": "jest --config ./jest.config.js",
-		"build": "rimraf dist && yarn run compile",
-		"compile": "tsc -p tsconfig.build.json"
+		"test:e2e": "jest --config ./jest.config.ts",
+		"build": "rimraf dist && tsc -p tsconfig.lib.json",
+		"build:prod": "rimraf dist && tsc -p tsconfig.lib.prod.json"
 	},
 	"peerDependencies": {
 		"tslib": "^2.6.2"
