+++ packages/plugins/jitsu-analytics/src/lib/jitsu-analytics.service.ts
