--- packages/plugins/jitsu-analytics/tsconfig.build.json
@@ -1,3 +1,0 @@
-{
-	"extends": "./tsconfig.json"
-}
