+++ packages/plugins/jitsu-analytics/tsconfig.lib.prod.json
@@ -1,0 +1,14 @@
+{
+	"extends": "./tsconfig.lib.json",
+	"compilerOptions": {
+		"declaration": true,
+		"sourceMap": false,
+		"removeComments": true,
+		"noEmitHelpers": true,
+		"importHelpers": true,
+		"target": "es6",
+		"outDir": "./dist"
+	},
+	"include": ["src/**/*.ts"],
+	"exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"]
+}
