+++ packages/plugins/jitsu-analytics/tsconfig.spec.json
@@ -1,15 +1,9 @@
 {
 	"extends": "./tsconfig.json",
 	"compilerOptions": {
-		"outDir": "../../dist/out-tsc",
+		"outDir": "../../../dist/out-tsc",
 		"module": "commonjs",
 		"types": ["jest", "node"]
 	},
-	"include": [
-		"**/*.spec.ts",
-		"**/*.spec.tsx",
-		"**/*.spec.js",
-		"**/*.spec.jsx",
-		"**/*.d.ts"
-	]
+	"include": ["jest.config.ts", "src/**/*.test.ts", "src/**/*.spec.ts", "src/**/*.d.ts"]
 }
